Added basic 2-player terminal game
When the user chooses to play the terminal game, the board state including the number of captures for each player will be output and the player will be asked for a row and column. If the move is invalid, the player will be prompted for another move until a valid move is entered. The player's piece will then be played at the corresponding position, with a red "O" as the player 1 token and a blue "O" as the player 2 token. All Pente rules are implemented: pairs of pieces can be captured by placing pieces either side of a pair of the opponent's pieces. The game continues until a 5-in-a-row or 5 captured pairs of pieces causes a player to win, or the board is full in which case it is a draw.