Added terminal account creation and login/logout mechanism, as well as input checking functions
A database holds the HashTable and Player tables. The HashTable table holds a hash table implemented as an array of empty lists, and when an account is created the username is appended to the list stored at the index of the player's password hash. A new player record is also created, storing the username and account creation datetime. The username is rejected by the login subroutine if it is not unique. When run a menu is displayed giving login, logout and account creation options. When playing a game, it also gives the option of letting the other player login. Two functions in the Terminal Ui getChoice (which accepts a number between two bounds) and getYesNo (which accepts either "y" or "n") enable inputs to be validated.