Added Game and PlayerGame tables and functions, and added GameRecord class
The Game and PlayerGame tables have been added and store information on the played games and the relations between the players and games respectively. Functions have been added to the Database program for loading, updating and saving records to the Game and PlayerGame tables. A new class called GameRecord has been added to the Game program. This acts as a record containing information on all the fields that is saved with a game. When loading games, all the game information is parsed and output as GameRecords, and when saving a game, all information is passed in as GameRecords.