Added Terminal Ui changes to enable players to load and save games
The current game record and current players are stored in the Terminal class. The menu for logged in players has been updated with an additional option to view games, after which the play can then load ongoing games. During a game, the player can quit, after which they have the choice of saving their game. If two players are logged in, the game will be saved to both accounts.