Added Ai program which uses minimax and alpha-beta pruning algorithm to play a move within reasonable time at a depth of 2, and altered Terminal Ui program to allow the different playing modes to be chosen
The Terminal Ui allows the user to choose between playing a 2-player game or a game against the computer. An Ai program has been added which uses the minimax algorithm combined with alpha-beta pruning to calculate the next move. Only moves next to pieces already on the board are considered, and if there are no such moves, the first empty space is chosen. The value of a board state is calculated by counting the number of "lines" of pieces, i.e. a space followed by n pieces of a colour followed by another space counts as one line of length n. The greater the length of the line, the more points are achieved. A line of 4 scores very highly as it almost guarantees a win for the player. Player 1 is the maximising player, and player 2 is the minimising player.