Added undo functionality using a move stack
Added a MoveStack class to the Game program. Each game has a MoveStack, and the captures, played row, and played column are pushed onto the stack after every move. When undoing, the row, column, and captures are popped off, and the captures are compared with that of the previous state. The Terminal Ui has been updated to give the player the option to undo.