Added subroutine which creates images of the empty cell and player pieces, and adjusted the buttons to vary in size depending on the board size
On initialisation the Gui class creates images using the PIL package of the empty cell and player 1 and 2 pieces, which can be overlayed onto the buttons to complete the overall board appearance. The buttons and image sizes depend on the square sizes (which is calculated from the canvas size and boardsize) to accomodate for changes in boardsize.