Players can click on intersections of the board and place pieces, playing a simple Pente game (capture count and winner displaying not included)
When the user clicks on an intersection (button) on the grid, if it is a valid move, the move will be played. The grid is then updated by iterating through each position in the board and updating the button image at each intersection accordingly. When the game ends, a winner (or draw) message is displayed in the terminal and no further moves can be played.