Added head label which displays the game mode and any error/end game messages. Added a play button. Made updating the board more efficient by only updating intersections which have changed.
A head label was added which displays the game mode, i.e. "Player v.s. Player", and any error messages and end-of-game messages. A play button was added. On startup, the canvas is blank. When pressing the play button, a grid appears and the user can play the game on the grid. The independent playGame subroutine adds the flexibility to adjust the board size (and therefore the square and canvas sizes) for different games. Before, placing a piece caused a slight lag as all the button images were updated regardless of changes made to the board. Now the new board is compared with the previous board, and only the necessary button images are updated, increasing efficiency.