Added a login button and an option frame
The main window is split into two frames: an option frame and a game frame. The option frames holds the play, clear and login buttons. On clicking the login button, a login window appears which gives the user two entry boxes to enter their username and password. On clicking the confirm button, if the username and password are valid the login window is destroyed. Otherwise, an error message appears at the bottom of the login window.