Added menuFrame and changed position and role of optionFrame. Added a quit button which appears when a game is being played. Created a logout button
The menuFrame holds login/logout buttons and the "Play new game" button. The optionFrame is situated to the right of the gameFrame, and holds options directly related to the currently being played game, e.g. "Quit". When a game is being played, the quit button will appear. On clicking it, a quit confirmation window will appear.