Changed currPlayers to store the Player.MAIN or Player.OPP rather than the actual players themselves. Instead of destroying the menu and option frames on update all the widgets are destroyed instead. Changed the name of the confirmLogin subroutine to login. Added self.playing variable. Removed the updateFrames subroutine. Added padding to the menu and option frame widgets.
Instead of storing Player.GUEST or the username of the player, currPlayers stores Player.MAIN and Player.OPP to mark which players are players 1 and 2. This is so that if the user logs in in the middle of the game, the game information will be updated of this change. Instead of destroying the menu and option frames on update, in order to make the update change less laggy on screen, the widgets are iterated through and independently destroyed. A self.playing variable was added, which stores a boolean value whether the user is currently playing a game or not.