Added create account functionality
If the user is not logged in, a create account button will be available. On clicking the button, a create account window appears and the user enters their username, password, and password confirmation. If the username is unique, the passwords match, and no entries are left blank then the new player record is saved and the user is automatically logged in. Otherwise, an error message will appear in the window.