Modified GameRecord so that it no longer has a winner property. Added save game functionality to the Gui.
The GameRecord no longer has a winner property, which was deemed redundant as it already stored the game which had the winner as an attribute. The Game table in the database still has a winner field. When the gui is run, if the user is logged in a save game button will be available. On clicking it, game information (players and game status) is displayed in a separate window and the user can enter a name to save the game as. On pressing confirm, providing the game name is not blank, a new record will be added to the database. Otherwise, an error message will be displayed in the save game window.