Added load game functionality to Gui. Added save game functionality for loaded games. Fixed bug where the button row and column values were swapped, meaning terminal game board was a reflection of the gui game board. Modified datetime string format so that month is shown first followed by the date. On running, the Gui creates the database tables.
When logged in, the user has the option to load a game which is shown in a separate window as a sequence of buttons. On clicking a game, the game is loaded onto the canvas. For loaded games, unlike new games, clicking the save game button will update the game in the database and no new window will appear.