Added choose game mode functionality (player vs player or player vs computer), with only player vs player implemented. Added option for other player to login when playing player vs player. The other player logging in is currently not required when loading games.
When selecting play new game, a window appears asking the user whether they want to play player vs player or player vs computer. On pressing player vs player it will ask whether the other player would like to log in. If not, a new game will start as normal. If they do, a login window will appear allowing the second player to log in. The second player logging in is not currently implemented for loading games.