Added randomisation of player number assignments to LAN play. Modified Game, Database and Ui programs to replace the 'computer' attribute of the GameRecord class to a more general 'mode' attribute.
Whenever two players are paired up to play each other via LAN, the computer randomly assigns which plays first. Instead of a 'computer' attribute of the GameRecord class which is set to True or False, there is now a 'mode' attribute which is set to "PVP", "LAN" or "COMP" to allow for the added LAN functionality.