Modified AI program to provide level 2 difficulty on top of 0 and 1
The getValue function was modified to improve the AI algorithm so that obvious moves to play weren't missed. This involved also creating two new functions getNumberOfCaptureLines and getNumberOfWinOpportunites to help calculate the value of a game state. The minimax function was edited so that only the game states with the top 3 values would be investigated further from each node. This was to increase the speed of the AI.