Added a function to the Database program which returns the number of games won, lost, drawn and ongoing by a player. Added an option to the terminal UI for users to view their profile.
On the terminal UI, users now have the option to view their profile which displays their username, the number of games they have played, won, lost, drawn, and have ongoing, in addition to the datetime the profile was created. This is done using a newly added function to the database. At the end of every game, the user (if it is not being played in player v.s. player LAN mode) is given the option to save the game.