Edited the AI getValue function to include values for winning game states, and modified the getNumberOfOpportunities function. Added an automatic save function which causes a game to be saved after ending.
Now the AI getValue function returns inf, -inf, or 0 for different winning situations if the game has ended. Otherwise it will calculate the value has normal. The getNumberOfWinOpportunities now, instead of comparing the number of capture lines with the number of existing captures, iterates through the board and finds the number of moves which the player can play to win via capturing and adds this to the number of win opportunities. A function was added so that if the user finishes playing the game, and at least one of the players are logged in (and the mode is not Player v.s. Player LAN), then the game is automatically saved.