Edited hard difficulty AI, fixed the comboBox bug in the Ui code
To decrease the time for the hard difficulty AI to come up with a move, the movesToAnalyse variable was changed from 3 to 2. For the UI, initially the starting value of the comboBox was passed in rather than the most recent value which resulted in the wrong games being loaded/deleted/exported. This bug is now fixed.